+++ SlashBar/App.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System.Windows;
 using SlashBar.Modules.Awake;
+using SlashBar.Modules.Pin;
 
 namespace SlashBar;
 
@@ -7,6 +8,7 @@ public partial class App : System.Windows.Application {
 
     protected override void OnExit(ExitEventArgs e) {
         AwakeSession.Disable();
+        PinSession.UnpinAll();
         base.OnExit(e);
     }
 }